work2

+++ backend/recipes/views.py
@@ -9,12 +9,12 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.viewsets import ReadOnlyModelViewSet
 
-from api.filters import AuthorAndTagFilter, IngredientSearchFilter
-from api.models import (Cart, Favorite, Ingredient, IngredientAmount, Recipe,
+from .filters import AuthorAndTagFilter, IngredientSearchFilter
+from .models import (ShopCart, Favorite, Ingredient, IngredientforRecipe, Recipe,
                         Tag)
-from api.pagination import LimitPageNumberPagination
-from api.permissions import IsAdminOrReadOnly, IsOwnerOrReadOnly, IsAuthorAdminOrReadOnly
-from api.serializers import (CropRecipeSerializer, IngredientSerializer,
+from .pagination import LimitPageNumberPagination
+from .permissions import IsAdminOrReadOnly, IsOwnerOrReadOnly, IsAuthorAdminOrReadOnly
+from .serializers import (IngredientSerializer,
                              RecipeSerializer, TagSerializer)
 
 
@@ -29,7 +29,6 @@ class IngredientsViewSet(ReadOnlyModelViewSet):
     queryset = Ingredient.objects.all()
     serializer_class = IngredientSerializer
     filter_backends = (IngredientSearchFilter,)
-    search_fields = ('^name',)
 
 
 class RecipeViewSet(viewsets.ModelViewSet):
@@ -59,20 +58,21 @@ class RecipeViewSet(viewsets.ModelViewSet):
                 status=status.HTTP_201_CREATED
         )
         if request.method == 'DELETE':
+            if favorite_recipe.exists():
                 favorite_recipe.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(detail=True, methods=['post', 'delete',],  url_path="shopping_cart")
     def shop_cart(self, request, pk):
         user = self.request.user
         recipe = get_object_or_404(Recipe, id=pk)
-        shopping_cart = Cart.objects.filter(recipe=recipe, user=user)
+        shopping_cart = ShopCart.objects.filter(recipe=recipe, user=user)
         if request.method == 'POST':
             if shopping_cart.exists():
                 content = {'status': 'Рецепт есть в списке покупок'}
                 return Response(content, status=status.HTTP_400_BAD_REQUEST)
         
-            Cart.objects.create(
+            ShopCart.objects.create(
                 user=self.request.user, recipe=recipe)
             return Response(
                 {'status': 'Рецепт добавлен в список покупок'},
@@ -84,55 +84,30 @@ class RecipeViewSet(viewsets.ModelViewSet):
                 return Response(status=status.HTTP_204_NO_CONTENT)
         return Response(status=status.HTTP_400_BAD_REQUEST)
 
-    @action(detail=False, methods=['get'])
-    def download_shopping_cart(self, request):
-        final_list = {}
-        ingredients = IngredientAmount.objects.filter(
-            recipe__cart__user=request.user).values_list(
-            'ingredient__name', 'ingredient__measurement_unit',
-            'amount')
-        for item in ingredients:
-            name = item[0]
-            if name not in final_list:
-                final_list[name] = {
-                    'measurement_unit': item[1],
-                    'amount': item[2]
-                }
-            else:
-                final_list[name]['amount'] += item[2]
-        pdfmetrics.registerFont(
-            TTFont('Slimamif', 'Slimamif.ttf', 'UTF-8'))
-        response = HttpResponse(content_type='application/pdf')
-        response['Content-Disposition'] = ('attachment; '
-                                           'filename="shopping_list.pdf"')
-        page = canvas.Canvas(response)
-        page.setFont('Slimamif', size=24)
-        page.drawString(200, 800, 'Список ингредиентов')
-        page.setFont('Slimamif', size=16)
-        height = 750
-        for i, (name, data) in enumerate(final_list.items(), 1):
-            page.drawString(75, height, (f'<{i}> {name} - {data["amount"]}, '
-                                         f'{data["measurement_unit"]}'))
-            height -= 25
-        page.showPage()
-        page.save()
+    
+    def get_shoppingcart(ist):
+        ingredients_dict = {}
+        for recipe in list:
+            ingredients = IngredientforRecipe.objects.filter(recipe=recipe.recipe)
+            for ingredient in ingredients:
+                amount = ingredient.amount
+                name = ingredient.ingredient.name
+                measurement_unit = ingredient.ingredient.measurement_unit
+                if name not in ingredients_dict:
+                    ingredients_dict[name] = {
+                        'measurement_unit': measurement_unit,
+                        'amount': amount
+                        }
+                else:
+                    ingredients_dict[name]['amount'] += amount
+        shop = []
+        for item in ingredients_dict:
+            shop.append(f'{item} - {ingredients_dict[item]["amount"]} '
+                        f'{ingredients_dict[item]["measurement_unit"]} \n')
+        return shop
+
+
+    def download(list, file):
+        response = HttpResponse(list, 'Content-Type: text/plain')
+        response['Content-Disposition'] = f'attachment; filen="{file}"'
         return response
-
-    def add_obj(self, model, user, pk):
-        if model.objects.filter(user=user, recipe__id=pk).exists():
-            return Response({
-                'errors': 'Рецепт уже добавлен в список'
-            }, status=status.HTTP_400_BAD_REQUEST)
-        recipe = get_object_or_404(Recipe, id=pk)
-        model.objects.create(user=user, recipe=recipe)
-        serializer = CropRecipeSerializer(recipe)
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-    def delete_obj(self, model, user, pk):
-        obj = model.objects.filter(user=user, recipe__id=pk)
-        if obj.exists():
-            obj.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        return Response({
-            'errors': 'Рецепт уже удален'
-        }, status=status.HTTP_400_BAD_REQUEST)
